backend/pyproject.toml
[tool.poetry]
name = "sem-art-search"
version = "0.1.0"
description = ""
authors = ["Sjors Lockhorst <sjors@lockhorst.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.18"
pgvector = "^0.2.5"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
transformers = "^4.41.2"
loguru = "^0.7.2"
pytest = "^8.2.2"
scikit-learn = "^1.5.0"
joblib = "^1.4.2"
matplotlib = "^3.9.0"
pydantic-settings = "^2.4.0"
torch = {version = "^2.4.1+cpu", source = "pytorch", optional = true}

[tool.poetry.extras]
torch = ["torch"]

[tool.poetry.group.backend]
optional = true

[tool.poetry.group.backend.dependencies]
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}

[tool.poetry.group.etl]
optional = true

[tool.poetry.group.etl.dependencies]
runpod = "^1.7.0"

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[tool.pylsp-mypy]
enabled = true
live_mode = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# match black
line-length = 120
target-version = "py312"
select = [
    "ALL", # include all the rules, including new ones
]
ignore = [
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports

    #### specific rules
    "D100",   # ignore missing docs
    "D101",
    "D102",
    "D103",
    "D104",
    "D105",
    "D106",
    "D107",
    "D200",
    "D205",
    "D212",
    "D400",
    "D401",
    "D415",
    "E402",   # false positives for local imports
    "E501",   # line too long
    "TRY003", # external messages in exceptions are too verbose
    "TD002",
    "TD003",
    "FIX002", # too verbose descriptions of todos
    "B008" # Not appliciable for fastapi app due to the Depends method
]
